Major Update
Added folders of whole_words training and validation data along with non working code to use with it. Added a few files of others code with the links attached which is required to make the non working code last a little longer. The non working code most likely has the opposite file access syntax as the OCR.py file since I was developing on windows but the OCR.py is set up to run on debian linux.